fix: Public Navbar Responsive
