+++ apps/frontend/src/app/Components/Header/HeaderInterfaces.tsx
@@ -1,0 +1,5 @@
+export interface NavItem {
+  label: string;
+  href?: string;
+  children?: NavItem[];
+}
